Pledges: Move `parse_contributors` function
