+++ plugins/wporg-5ftf/includes/contributor.php
@@ -402,3 +402,58 @@ function process_my_pledges_form() {
 
 	return $message;
 }
+
+/**
+ * Ensure each item in a list of usernames is valid and corresponds to a user.
+ *
+ * @param string $contributors A comma-separated list of username strings.
+ *
+ * @return array|WP_Error An array of sanitized wporg usernames on success. Otherwise WP_Error.
+ */
+function parse_contributors( $contributors ) {
+	$invalid_contributors   = array();
+	$sanitized_contributors = array();
+
+	$contributors = str_replace( '@', '', $contributors );
+	$contributors = explode( ',', $contributors );
+
+	foreach ( $contributors as $wporg_username ) {
+		$sanitized_username = sanitize_user( $wporg_username );
+		$user               = get_user_by( 'login', $sanitized_username );
+
+		if ( ! $user instanceof WP_User ) {
+			$user = get_user_by( 'slug', $sanitized_username );
+		}
+
+		if ( $user instanceof WP_User ) {
+			$sanitized_contributors[] = $user->user_login;
+		} else {
+			$invalid_contributors[] = $wporg_username;
+		}
+	}
+
+	if ( ! empty( $invalid_contributors ) ) {
+		/* translators: Used between sponsor names in a list, there is a space after the comma. */
+		$item_separator = _x( ', ', 'list item separator', 'wporg-5ftf' );
+
+		return new WP_Error(
+			'invalid_contributor',
+			sprintf(
+				/* translators: %s is a list of usernames. */
+				__( 'The following contributor usernames are not valid: %s', 'wporg-5ftf' ),
+				implode( $item_separator, $invalid_contributors )
+			)
+		);
+	}
+
+	if ( empty( $sanitized_contributors ) ) {
+		return new WP_Error(
+			'contributor_required',
+			__( 'The pledge must have at least one contributor username.', 'wporg-5ftf' )
+		);
+	}
+
+	$sanitized_contributors = array_unique( $sanitized_contributors );
+
+	return $sanitized_contributors;
+}
